+++ utils/automatic_saving.bash
@@ -7,6 +7,33 @@ DEFAULT_INTERVAL_TIME=10
 readonly I3_RESTORE_SAVE_FILE DEFAULT_INTERVAL_TIME
 
 #####################################
+# Get the sleep time (in seconds) from the
+# argument. If the argument is not a number,
+# 0, or not provided, the default time will be
+# returned
+# Globals:
+#   DEFAULT_INTERVAL_TIME
+# Arguments:
+#   sleep time, in minutes
+# Returns:
+#   the sleep time in seconds
+#####################################
+get_sleep_time() {
+    local sleep_time="${1}"
+
+    if ! [[ ${sleep_time} =~ ^[0-9]+$ ]] || [[ ${sleep_time} == 0 ]]; then
+        message="Sleep time not passed in (or is invalid). Using "
+        message+="default time of ${DEFAULT_INTERVAL_TIME} minutes"
+        log "${message}"
+        sleep_time="${DEFAULT_INTERVAL_TIME}"
+    fi
+
+    # Convert sleep time to seconds
+    sleep_time=$((sleep_time * 60))
+    echo "${sleep_time}"
+}
+
+#####################################
 # Ensure the same i3 process that
 # automatic saving was started with
 # is still running. Otherwise, i3-msg
@@ -32,18 +59,9 @@ check_i3_alive() {
 #   I3_RESTORE_INTERVAL_MINUTES
 #####################################
 start_save_interval() {
-    local i3_pid
-    local sleep_time="${I3_RESTORE_INTERVAL_MINUTES}"
+    local i3_pid sleep_time
 
-    # Ensure sleep time is passed in and is a number, otherwise use default interval
-    if ! [[ ${sleep_time} =~ ^[0-9]+$ ]] || [[ ${sleep_time} == 0 ]]; then
-        local message="Sleep time not passed in (or is invalid). Using "
-        message+="default time of ${DEFAULT_INTERVAL_TIME} minutes"
-        log "${message}"
-        sleep_time="${DEFAULT_INTERVAL_TIME}"
-    fi
-
-    sleep_time="$((sleep_time * 60))"
+    sleep_time="$(get_sleep_time "${I3_RESTORE_INTERVAL_MINUTES}")"
     i3_pid="$(pidof i3)"
 
     log "Starting automatic saving on an interval of ${sleep_time} seconds"