+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 When upgrading to a new version, make sure to follow the directions under the "Upgrading" header of the corresponding version.
 If there is no "Upgrading" header for that version, no post-upgrade actions need to be performed.
 
+## Upcoming
+
+### Bug Fixes
+- Fix log messages messing up the restoring process when using a verbosity
+
 ## 4.0 (2024-01-03)
 
 ### New Features