+++ CHANGELOG.md
@@ -6,6 +6,7 @@ If there is no "Upgrading" header for that version, no post-upgrade actions need
 
 ### Bug Fixes
 - Fix log messages messing up the restoring process when using a verbosity
+- Fix programs not restoring correctly when the default shell is not Bash
 
 ## 4.0 (2024-01-03)